+++ src/store/modules/connext.ts
@@ -225,7 +225,27 @@ const actions = <ActionTree<ConnextState, RootState>>{
     } else {
       console.log('not ready to claim')
     }
-  }
+  },
+
+  async cancelTransfer({ dispatch, rootState }, activeTransaction: ActiveTransaction) {
+    const { sending, invariant } = activeTransaction.crosschainTx;
+    const sendingTxData = {
+      ...invariant,
+      ...sending,
+    }
+
+    if (!connextSDK) {
+      await dispatch('instantiateConnext')
+    }
+    if (!rootState.wallet.connected) {
+      await dispatch('connectWallet')
+    }
+
+    await connextSDK.cancel(
+      { signature: "0x", txData: sendingTxData },
+      activeTransaction.crosschainTx.invariant.sendingChainId
+    )
+  },
 }
 
 const getters = <GetterTree<ConnextState, RootState>>{