+++ CHANGELOG.md
@@ -11,3 +11,7 @@ Unreleased
 - adjusts time remaining logic
 - style footer logos
 - fix unwanted caching by removing service worker
+- show more useful error when entering an amount without an asset selected
+- split up `Couldn't setup connext` errors to better debug
+- changes settings to dropdown, adds "Search transaction" link
+- removes "Cancel" button for Connext transactions, expired txs can be handled in Connextscan